isbn-ocr/requirements.txt
flask
pytesseract
opencv-python
Pillow